PostImage now shows imgix images if possible. Made maxAutoHeight different based on device type.
